+++ sipXregistry/src/SipRegistrarServer.h
@@ -112,7 +112,7 @@ protected:
                             *   which contains the AOR to register to. */
         const UtlString& instrument,
                            ///< instrument value to add to bindings
-        const int timeNow, ///< base time for all expiration calculations
+        const unsigned long timeNow, ///< base time for all expiration calculations
         const SipMessage& registerMessage, ///< message containing bindings
         RegistrationExpiryIntervals*& expiryIntervalsUsed ); ///< returns the expiry interval used to bound the expiry of the registration
